# src/ViajantesTurismo.Admin.Contracts.Http/OpenApi/v1.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ViajantesTurismo.Admin.ApiService | v1",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/bookings": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Retrieves all bookings.",
        "description": "Retrieves all bookings.",
        "operationId": "GetBookings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetBookingDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The authenticated caller lacks the required permission."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Creates a new booking.",
        "description": "Creates a new booking for a tour.",
        "operationId": "CreateBooking",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBookingDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The authenticated caller lacks the required permission."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/bookings/{id}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Retrieves a booking by its ID.",
        "description": "Retrieves a booking by its ID.",
        "operationId": "GetBookingById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBookingDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Bookings"
        ],
        "summary": "Deletes a booking.",
        "description": "Deletes a booking.",
        "operationId": "DeleteBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings/tour/{tourId}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Retrieves all bookings for a specific tour.",
        "description": "Retrieves all bookings for a specific tour.",
        "operationId": "GetBookingsByTourId",
        "parameters": [
          {
            "name": "tourId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetBookingDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings/customer/{customerId}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Retrieves all bookings for a specific customer.",
        "description": "Retrieves all bookings for a specific customer (as primary or companion).",
        "operationId": "GetBookingsByCustomerId",
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetBookingDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/bookings/{id}/confirm": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Confirms a booking.",
        "description": "Confirms a booking by transitioning its status to Confirmed.",
        "operationId": "ConfirmBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBookingDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/bookings/{id}/payments": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Records a payment.",
        "description": "Records a payment for a booking.",
        "operationId": "RecordPayment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPaymentDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BikeTypeDto": {
        "type": "integer"
      },
      "BookingStatusDto": {
        "type": "integer"
      },
...
      "CreateBookingDto": {
        "required": [
          "tourId",
          "principalCustomerId",
          "principalBikeType",
          "roomType"
        ],
        "type": "object",
        "properties": {
          "tourId": {
            "type": "string",
            "format": "uuid"
          },
          "principalCustomerId": {
            "type": "string",
            "format": "uuid"
          },
          "principalBikeType": {
            "$ref": "#/components/schemas/BikeTypeDto"
          },
          "companionCustomerId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "companionBikeType": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BikeTypeDto"
              }
            ]
          },
          "roomType": {
            "$ref": "#/components/schemas/RoomTypeDto"
          },
          "discountType": {
            "$ref": "#/components/schemas/DiscountTypeDto"
          },
          "discountAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "discountReason": {
            "maxLength": 500,
            "minLength": 10,
            "type": [
              "null",
              "string"
            ]
          },
          "notes": {
            "maxLength": 2000,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CreatePaymentDto": {
        "required": [
          "amount",
          "paymentDate",
          "method"
        ],
        "type": "object",
        "properties": {
          "amount": {
            "minimum": 0.01,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "paymentDate": {
            "type": "string",
            "format": "date-time"
          },
          "method": {
            "$ref": "#/components/schemas/PaymentMethodDto"
          },
          "referenceNumber": {
            "maxLength": 128,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "notes": {
            "maxLength": 500,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CurrencyDto": {
        "type": "integer"
      },
...
      "DiscountTypeDto": {
        "type": "integer"
      },
...
      "GetBookingDto": {
        "required": [
          "id",
          "tourId",
          "tourIdentifier",
          "tourName",
          "customerId",
          "customerName",
          "bookingDate",
          "status",
          "paymentStatus",
          "totalPrice",
          "discountType",
          "discountAmount",
          "currency",
          "payments",
          "amountPaid",
          "remainingBalance"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tourId": {
            "type": "string",
            "format": "uuid"
          },
          "tourIdentifier": {
            "type": "string"
          },
          "tourName": {
            "type": "string"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "customerName": {
            "type": "string"
          },
          "companionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "companionName": {
            "type": [
              "null",
              "string"
            ]
          },
          "bookingDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatusDto"
          },
          "paymentStatus": {
            "$ref": "#/components/schemas/PaymentStatusDto"
          },
          "totalPrice": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "discountType": {
            "$ref": "#/components/schemas/DiscountTypeDto"
          },
          "discountAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "discountReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyDto"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetPaymentDto"
            }
          },
          "amountPaid": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "remainingBalance": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "GetPaymentDto": {
        "required": [
          "id",
          "bookingId",
          "amount",
          "paymentDate",
          "method",
          "recordedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "bookingId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "paymentDate": {
            "type": "string",
            "format": "date-time"
          },
          "method": {
            "$ref": "#/components/schemas/PaymentMethodDto"
          },
          "referenceNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "PaymentMethodDto": {
        "type": "integer"
      },
      "PaymentStatusDto": {
        "type": "integer"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RoomTypeDto": {
        "type": "integer"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}